+++ src/XRP/issued-currency-client.ts
@@ -1,19 +1,29 @@
-import { Wallet, XrplNetwork } from 'xpring-common-js'
-import GrpcNetworkClient from './network-clients/grpc-xrp-network-client'
-import GrpcNetworkClientWeb from './network-clients/grpc-xrp-network-client.web'
-import { XrpNetworkClient } from './network-clients/xrp-network-client'
-import isNode from '../Common/utils'
-import CoreXrplClient from './core-xrpl-client'
-import { LimitAmount } from './Generated/web/org/xrpl/rpc/v1/common_pb'
-import XrpUtils from './shared/xrp-utils'
-import { TransactionResult, XrpError } from './shared'
+import { XrplNetwork, XrpUtils, Wallet } from 'xpring-common-js'
+
+import { LimitAmount, SetFlag } from './Generated/web/org/xrpl/rpc/v1/common_pb'
 import { AccountAddress } from './Generated/web/org/xrpl/rpc/v1/account_pb'
 import {
   Currency,
   CurrencyAmount,
   IssuedCurrencyAmount,
 } from './Generated/web/org/xrpl/rpc/v1/amount_pb'
-import { TrustSet } from './Generated/web/org/xrpl/rpc/v1/transaction_pb'
+import {
+  TrustSet,
+  AccountSet,
+} from './Generated/web/org/xrpl/rpc/v1/transaction_pb'
+
+import GrpcNetworkClient from './network-clients/grpc-xrp-network-client'
+import GrpcNetworkClientWeb from './network-clients/grpc-xrp-network-client.web'
+import { GrpcNetworkClientInterface } from './network-clients/grpc-network-client-interface'
+import JsonRpcNetworkClient from './network-clients/json-rpc-network-client'
+import isNode from '../Common/utils'
+import CoreXrplClient from './core-xrpl-client'
+import { XrpError } from './shared'
+import { AccountSetFlag } from './shared/account-set-flag'
+import TransactionResult from './shared/transaction-result'
+import TrustLine from './shared/trustline'
+import { AccountLinesResponse } from './shared/rippled-json-rpc-schema'
+import { JsonNetworkClientInterface } from './network-clients/json-network-client-interface'
 
 /**
  * IssuedCurrencyClient is a client for working with Issued Currencies on the XRPL.
@@ -33,12 +43,19 @@ export default class IssuedCurrencyClient {
    */
   public static issuedCurrencyClientWithEndpoint(
     grpcUrl: string,
+    jsonUrl: string,
     network: XrplNetwork,
     forceWeb = false,
   ): IssuedCurrencyClient {
-    return isNode() && !forceWeb
-      ? new IssuedCurrencyClient(new GrpcNetworkClient(grpcUrl), network)
-      : new IssuedCurrencyClient(new GrpcNetworkClientWeb(grpcUrl), network)
+    const grpcNetworkClient =
+      isNode() && !forceWeb
+        ? new GrpcNetworkClient(grpcUrl)
+        : new GrpcNetworkClientWeb(grpcUrl)
+    return new IssuedCurrencyClient(
+      grpcNetworkClient,
+      new JsonRpcNetworkClient(jsonUrl),
+      network,
+    )
   }
 
   /**
@@ -50,14 +67,11 @@ export default class IssuedCurrencyClient {
    * @param network The network this IssuedCurrencyClient is connecting to.
    */
   public constructor(
-    private readonly networkClient: XrpNetworkClient,
+    grpcNetworkClient: GrpcNetworkClientInterface,
+    private readonly jsonNetworkClient: JsonNetworkClientInterface,
     readonly network: XrplNetwork,
   ) {
-    this.coreXrplClient = new CoreXrplClient(networkClient, network)
-  }
-
-  public shutUpCompiler(): void {
-    console.log(this.networkClient)
+    this.coreXrplClient = new CoreXrplClient(grpcNetworkClient, network)
   }
 
   /**
@@ -111,6 +125,70 @@ export default class IssuedCurrencyClient {
       wallet,
     )
 
-    return await this.coreXrplClient.getTransactionResult(transactionHash)
+    return await this.coreXrplClient.getFinalTransactionResultAsync(
+      transactionHash,
+      wallet,
+    )
+  }
+  /**
+   * Enable Require Authorization for this XRPL account.
+   *
+   * @see https://xrpl.org/become-an-xrp-ledger-gateway.html#require-auth
+   *
+   * @param wallet The wallet associated with the XRPL account enabling Require Authorization and that will sign the request.
+   * @returns A promise which resolves to a TransactionResult object that represents the result of this transaction.
+   */
+  public async requireAuthorizedTrustlines(
+    wallet: Wallet,
+  ): Promise<TransactionResult> {
+    const setFlag = new SetFlag()
+    setFlag.setValue(AccountSetFlag.asfRequireAuth)
+
+    const accountSet = new AccountSet()
+    accountSet.setSetFlag(setFlag)
+
+    const transaction = await this.coreXrplClient.prepareBaseTransaction(wallet)
+    transaction.setAccountSet(accountSet)
+
+    const transactionHash = await this.coreXrplClient.signAndSubmitTransaction(
+      transaction,
+      wallet,
+    )
+
+    return await this.coreXrplClient.getFinalTransactionResultAsync(
+      transactionHash,
+      wallet,
+    )
+  }
+
+  /**
+   * Retrieves information about an account's trust lines, which maintain balances of all non-XRP currencies and assets.
+   * @see https://xrpl.org/trust-lines-and-issuing.html
+   *
+   * @param account The account for which to retrieve associated trust lines, encoded as an X-Address.
+   * @see https://xrpaddress.info/
+   * @returns An array of TrustLine objects, representing all trust lines associated with this account.
+   */
+  public async getTrustLines(account: string): Promise<Array<TrustLine>> {
+    const classicAddress = XrpUtils.decodeXAddress(account)
+    if (!classicAddress) {
+      throw XrpError.xAddressRequired
+    }
+    const accountLinesResponse: AccountLinesResponse = await this.jsonNetworkClient.getAccountLines(
+      classicAddress.address,
+    )
+
+    if (accountLinesResponse.result.error) {
+      throw XrpError.accountNotFound
+    }
+    const rawTrustLines = accountLinesResponse.result.lines
+    if (rawTrustLines === undefined) {
+      throw XrpError.malformedResponse
+    }
+    const trustLines: Array<TrustLine> = []
+    rawTrustLines.map((trustLineJson) => {
+      trustLines.push(new TrustLine(trustLineJson))
+    })
+    return trustLines
   }
 }