+++ src/XRP/network-clients/json-network-client-interface.ts
@@ -1,0 +1,8 @@
+import { AccountLinesResponse } from '../shared/rippled-json-rpc-schema'
+
+/**
+ * The JsonNetworkClientInterface defines the calls available via the rippled JSON-RPC.
+ */
+export interface JsonNetworkClientInterface {
+  getAccountLines(account: string): Promise<AccountLinesResponse>
+}