+++ src/XRP/xrp-client-decorator.ts
@@ -111,4 +111,17 @@ export default interface XrpClientDecorator {
     xAddressToAuthorize: string,
     wallet: Wallet,
   ): Promise<TransactionResult>
+
+  /**
+   * Unauthorizes an XRPL account to send to this XRPL account.
+   *
+   * @see https://xrpl.org/depositpreauth.html
+   *
+   * @param xAddressToUnauthorize The X-Address of the account to unauthorize as a sender.
+   * @param wallet The wallet associated with the XRPL account unauthorizing a sender, and that will sign the request.
+   */
+  unauthorizeSendingAccount(
+    xAddressToUnauthorize: string,
+    wallet: Wallet,
+  ): Promise<TransactionResult>
 }